requirements.txt
arabic-reshaper==2.1.3
asgiref==3.6.0
asn1crypto==1.5.1
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
boto3==1.28.32
botocore==1.31.32
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
cloudinary==1.29.0
colorama==0.4.5
command-not-found==0.3
crashtest==0.4.1
crispy-bootstrap5==0.7
cryptography==39.0.2
cssselect2==0.6.0
dbus-python==1.2.16
defusedxml==0.7.1
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==2.1.0
Django==4.2
django-allauth==0.54.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-cors-headers==4.2.0
django-countries==7.4.2
django-crispy-forms==2.0
django-daraja==1.2.0
django-decouple==2.1
django-environ==0.10.0
django-flexible-subscriptions==0.15.1
django-formtools==2.4
django-fsm==2.8.1
django-fsm-log==3.1.0
django-phonenumber-field==6.3.0
django-pwa==1.1.0
django-storages==1.13.2
django-tinymce==3.6.1
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
drf-spectacular==0.26.4
dulwich==0.21.3
environ==1.0
filelock==3.9.0
future==0.18.2
gunicorn==21.2.0
html5lib==1.1
httpie==3.2.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.0.0
inflection==0.5.1
installer==0.6.0
izitoast==1.2.2
jaraco.classes==3.2.3
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.17.3
keyring==23.13.1
language-selector==0.1
lockfile==0.12.2
lxml==4.9.2
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
netifaces==0.10.4
nine==1.1.0
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
oscrypto==1.3.0
packaging==23.0
pexpect==4.8.0
phonenumberslite==8.12.51
Pillow==9.5.0
pipenv==2023.2.18
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.15.1
PyGObject==3.36.0
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
PyPDF3==1.0.6
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-apt==2.0.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-utils==3.6.0
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyuca==1.2
PyYAML==5.3.1
qrcode==7.3.1
rapidfuzz==2.13.7
reportlab==3.6.9
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.5.2
s3transfer==0.6.2
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.4
ssh-import-id==5.10
supervisor==4.1.0
svglib==1.5.1
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.0
trove-classifiers==2023.3.9
typing_extensions==4.6.3
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.12
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.15
utils==1.0.1
validate-email==1.3
virtualenv==20.20.0
virtualenv-clone==0.5.7
webencodings==0.5.1
whitenoise==6.4.0
xhtml2pdf==0.2.7
zipp==3.15.0
